webhook 3

+++ functions/routes/webhooks/redeemPoints.js
@@ -15,7 +15,8 @@ exports.post = ({ appSdk }, req, res) => {
    * Ref.: https://developers.e-com.plus/docs/api/#/store/triggers/
    */
   const trigger = req.body
-console.log(JSON.stringify(trigger))
+  const resourceId = trigger.resource_id || trigger.inserted_id
+  //console.log(JSON.stringify(trigger))
   // get app configured options
   getAppData({ appSdk, storeId })
 
@@ -32,37 +33,52 @@ console.log(JSON.stringify(trigger))
 
       /* DO YOUR CUSTOM STUFF HERE */
 
-      if (appData.instancia && trigger.resource === 'orders') {      
-        admin.firestore().doc(`prizes/${storeId}_${params.customer._id}`).get()
-        .then(function(result){
-          const reg = result.data()
-          let prize_id = reg.selected_prize_id
-          if (reg.selected_prize_id) {
-            const docNumber = reg.doc_number
-            const crmUrl = `${appData.instancia}/cgi-bin/webworks/bin/sharkview_api_v1?id=${appData.id}&token=${appData.token}&cmd=points_redemption&cpf=${docNumber}&order=${order}&id_prize=${prize_id}`
-            axios.get(crmUrl)
-              .then(({ data }) => {
-                admin.firestore().doc(`prizes/${storeId}_${params.customer._id}`).delete()
-                .then(function(){
-                  res.status(200).send({
-                    prize: params.customer._id + ' - ' + prize_id,
-                    message: 'success'
+      if (appData.instancia && trigger.resource === 'orders' && trigger.action === 'create') {   
+        appSdk.getAuth(storeId).then( authorization => {
+          appSdk.apiRequest(storeId, `orders/${resourceId}.json`, 'GET', null, authorization)
+          .then(({ order }) => {
+            admin.firestore().doc(`prizes/${storeId}_${params.customer._id}`).get()
+            .then(function(result){
+              const reg = result.data()
+              let prize_id = reg.selected_prize_id
+              if (reg.selected_prize_id) {
+                const docNumber = reg.doc_number
+                const crmUrl = `${appData.instancia}/cgi-bin/webworks/bin/sharkview_api_v1?id=${appData.id}&token=${appData.token}&cmd=points_redemption&cpf=${docNumber}&order=${order}&id_prize=${prize_id}`
+                axios.get(crmUrl)
+                  .then(({ data }) => {
+                    admin.firestore().doc(`prizes/${storeId}_${params.customer._id}`).delete()
+                    .then(function(){
+                      res.status(200).send({
+                        prize: params.customer._id + ' - ' + prize_id,
+                        message: 'success'
+                      })
+                    })
                   })
-                })
-              })
-              .catch(err => {
-                console.log(JSON.stringify({
-                  crmUrl,
-                  resStatus: err.response?.status,
-                  resData: err.response?.data
-                }))
-                res.status(409).send({
-                  error: '1',
-                  message: err.message
-                })
-              })
-          }      
+                  .catch(err => {
+                    console.log(JSON.stringify({
+                      crmUrl,
+                      resStatus: err.response?.status,
+                      resData: err.response?.data
+                    }))
+                    res.status(409).send({
+                      error: '1',
+                      message: err.message
+                    })
+                  })
+              }      
+            })
+          })
+          .catch((err) => {
+            res.status(500)
+            const { message } = err
+            res.send({
+              error: ECHO_API_ERROR,
+              message
+            })
+          }) 
         })
+          
+        
       }
 
       // all done
